re structured files

+++ routes/web.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Http\Controllers\ProfileController;
+use App\Http\Controllers\Customer\ProfileController;
 use Illuminate\Support\Facades\Route;
 use Inertia\Inertia;
 
@@ -22,44 +22,19 @@ Route::get('/', function () {
     ]);
 })->name('home');
 
-Route::get('/dashboard', function () {
-    return Inertia::render('Dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
 
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
-
-    Route::resource('products', \App\Http\Controllers\ProductController::class);
-    Route::post('products/get_slug', [\App\Http\Controllers\ProductController::class, 'get_slug'])
-        ->name('products.get_slug');
-    Route::delete('products/{product}/{media}', [\App\Http\Controllers\ProductController::class, 'deleteMedia'])->name('products.deleteMedia');
-
-    Route::resource('categories', \App\Http\Controllers\CategoryController::class)
-        ->except('show');
-    Route::post('categories/get_names', [\App\Http\Controllers\CategoryController::class, 'get_names'])
-        ->name('categories.get_names');
-    Route::post('categories/get_slug', [\App\Http\Controllers\CategoryController::class, 'get_slug'])
-        ->name('categories.get_slug');
-    Route::delete('categories/{category}/{media}', [\App\Http\Controllers\CategoryController::class, 'deleteMedia'])
-        ->name('categories.deleteMedia');
-
-    Route::resource('brands', \App\Http\Controllers\BrandController::class);
-    Route::post('brands/get_names', [\App\Http\Controllers\BrandController::class, 'get_names'])
-        ->name('brands.get_names');
-    Route::post('brands/get_slug', [\App\Http\Controllers\BrandController::class, 'get_slug'])
-        ->name('brands.get_slug');
-
-    Route::resource('users', \App\Http\Controllers\UserController::class);
 });
 
 /**
  * Cart routes
  */
-Route::post('cart/{product}', [\App\Http\Controllers\CartController::class, 'store'])
+Route::post('cart/{product}', [\App\Http\Controllers\Customer\CartController::class, 'store'])
     ->name('cart.store');
-Route::delete('cart/{rowId}', [\App\Http\Controllers\CartController::class, 'destroy'])
+Route::delete('cart/{rowId}', [\App\Http\Controllers\Customer\CartController::class, 'destroy'])
     ->name('cart.delete');
 Route::get('cart', function () {
     return Inertia::render('Customer/Cart', [
@@ -78,20 +53,20 @@ Route::middleware('auth')->group(function () {
             'products' => \App\Http\Resources\CartItemResource::collection(Cart::instance('wishlist')->content()),
         ]);
     })->name('wishlist');
-    Route::post('wishlist/{product}', [\App\Http\Controllers\WishlistController::class, 'store'])->name('wishlist.store');
-    Route::delete('wishlist/{rowId}', [\App\Http\Controllers\WishlistController::class, 'destroy'])->name('wishlist.delete');
+    Route::post('wishlist/{product}', [\App\Http\Controllers\Customer\WishlistController::class, 'store'])->name('wishlist.store');
+    Route::delete('wishlist/{rowId}', [\App\Http\Controllers\Customer\WishlistController::class, 'destroy'])->name('wishlist.delete');
 
     /**
      * checkout routes
      */
 
-    Route::post('checkout', [\App\Http\Controllers\CheckoutController::class, 'store'])->name('checkout.store');
-    Route::get('checkout', [\App\Http\Controllers\CheckoutController::class, 'show'])->name('checkout.show');
+    Route::post('checkout', [\App\Http\Controllers\Customer\CheckoutController::class, 'store'])->name('checkout.store');
+    Route::get('checkout', [\App\Http\Controllers\Customer\CheckoutController::class, 'show'])->name('checkout.show');
 
 
-    Route::get('orders', [\App\Http\Controllers\OrderController::class, 'index'])->name('orders.index');
-    Route::get('orders/{order}', [\App\Http\Controllers\OrderController::class, 'show'])->name('orders.show');
-    Route::delete('orders/{order}', [\App\Http\Controllers\OrderController::class, 'destroy'])->name('orders.destroy');
+    Route::get('orders', [\App\Http\Controllers\Admin\OrderController::class, 'index'])->name('orders.index');
+    Route::get('orders/{order}', [\App\Http\Controllers\Admin\OrderController::class, 'show'])->name('orders.show');
+    Route::delete('orders/{order}', [\App\Http\Controllers\Admin\OrderController::class, 'destroy'])->name('orders.destroy');
 
     /**
      * profile routes
@@ -99,22 +74,14 @@ Route::middleware('auth')->group(function () {
     Route::get('profile', function () {
         return Inertia::render('Customer/Profile');
     })->name('profile');
-    Route::post('profile', [\App\Http\Controllers\ProfileController::class, 'update'])->name('profile.update');
+    Route::post('profile', [\App\Http\Controllers\Customer\ProfileController::class, 'update'])->name('profile.update');
     Route::get('address', [ProfileController::class, 'editAddress'])->name('address');
-    Route::post('address', [\App\Http\Controllers\ProfileController::class, 'updateAddress'])->name('address.store');
+    Route::post('address', [\App\Http\Controllers\Customer\ProfileController::class, 'updateAddress'])->name('address.store');
     Route::get('password', [ProfileController::class, 'editPassword'])->name('password');
-    Route::get('reviews', [\App\Http\Controllers\ProfileController::class, 'showReviews'])->name('reviews.index');
+    Route::get('reviews', [\App\Http\Controllers\Customer\ProfileController::class, 'showReviews'])->name('reviews.index');
 
-    Route::resource('review', \App\Http\Controllers\ReviewController::class)->only('store');
+    Route::resource('review', \App\Http\Controllers\Customer\ReviewController::class)->only('store');
 });
-
-Route::get('login', function () {
-    return Inertia::render('Customer/Login');
-})->name('login');
-
-Route::get('register', function () {
-    return Inertia::render('Customer/Register');
-})->name('register');
 
 Route::get('icons', function () {
     return Inertia::render('Icons');
